+++ src/extension.ts
@@ -11,6 +11,7 @@ export function activate(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand("listen.refreshAllPodcasts", listen.libraryProvider.updateAllPodcasts),
         vscode.commands.registerCommand("listen.refreshPodcast", listen.libraryProvider.updatePodcast),
         vscode.commands.registerCommand("listen.markPodcastAsListened", listen.libraryProvider.markPodcastAsListened),
+        vscode.commands.registerCommand("listen.markEpisodeAsListened", listen.libraryProvider.markEpisodeAsListened),
         vscode.commands.registerCommand("listen.removePodcast", listen.libraryProvider.removePodcast),
         vscode.commands.registerCommand("listen.addRadioStream", listen.radio.openDialog),
         vscode.commands.registerCommand("listen.removeRadio", listen.radio.remove),